Rename clean_noise to clean_flicker_noise
