+++ jwst/clean_flicker_noise/__init__.py
@@ -1,0 +1,3 @@
+from .clean_flicker_noise_step import CleanFlickerNoiseStep
+
+__all__ = ['CleanFlickerNoiseStep']
