--- jwst/clean_noise/__init__.py
@@ -1,3 +1,0 @@
-from .clean_noise_step import CleanNoiseStep
-
-__all__ = ['CleanNoiseStep']
