+++ jwst/lib/suffix.py
@@ -159,8 +159,8 @@ _calculated_suffixes = {
  'spec3pipeline',
  'nsclean',
  'nscleanstep',
- 'clean_noise',
- 'cleannoisestep',
+ 'clean_flicker_noise',
+ 'cleanflickernoisestep',
  'outlierdetectionstep',
  'groupscalestep',
  'hlspstep',
