+++ jwst/pipeline/calwebb_detector1.py
@@ -22,7 +22,7 @@ from ..reset import reset_step
 from ..persistence import persistence_step
 from ..charge_migration import charge_migration_step
 from ..jump import jump_step
-from ..clean_noise import clean_noise_step
+from ..clean_flicker_noise import clean_flicker_noise_step
 from ..ramp_fitting import ramp_fit_step
 from ..gain_scale import gain_scale_step
 
@@ -65,7 +65,7 @@ class Detector1Pipeline(Pipeline):
                  'persistence': persistence_step.PersistenceStep,
                  'charge_migration': charge_migration_step.ChargeMigrationStep,
                  'jump': jump_step.JumpStep,
-                 'clean_noise': clean_noise_step.CleanNoiseStep,
+                 'clean_flicker_noise': clean_flicker_noise_step.CleanFlickerNoiseStep,
                  'ramp_fit': ramp_fit_step.RampFitStep,
                  'gain_scale': gain_scale_step.GainScaleStep,
                  }
@@ -130,8 +130,8 @@ class Detector1Pipeline(Pipeline):
         # apply the jump step
         input = self.jump(input)
 
-        # apply the clean_noise step
-        input = self.clean_noise(input)
+        # apply the clean_flicker_noise step
+        input = self.clean_flicker_noise(input)
 
         # save the corrected ramp data, if requested
         if self.save_calibrated_ramp:
