+++ jwst/pipeline/clean_noise.cfg
@@ -1,2 +1,2 @@
-name = "clean_noise"
-class = "jwst.clean_noise.CleanNoiseStep"
+name = "clean_flicker_noise"
+class = "jwst.clean_flicker_noise.CleanFlickerNoiseStep"
